(refactoring) using context.WithTimeout instead of time.NewTimer
We noticed that work with time precision is hard with
parallel/concurrent program. Golang offer a pattern to deal with timeout
that uses context.WithTimeout function and we are using it now.

Besides that, we increase the gap between the time used in the test
called
"TestGrelayRequestExecWithOneItemInQueueReturningErrGrelayServiceTimedoutShouldReturnErrGrelayServiceTimedout". Because there is no documented guaranty how precise time is. We must not assume some (especially low) precision.